Allow control over loader `exec`.
An `exec` flag can be provided to the loader query or webpack config. This instructs the loader to load the source using `exec`. This enhances the previous behavior which was to check that the loader query parser was strictly equal to `postcss-js`.

This fixes an issue where the `postcss-js` parser cannot be specified through the webpack postcss config option.

This also enables parsers other than `postcss-js` to receive executed module results.